Refactored build stylesheets: Changed file naming convention to TitleCase. Split out CSharp codegen stylesheet into multiple stylesheet to make it easier to understand.
Improved the build batch files:
Added error checking and handling to batch scripts.
Added filetime comparisons to detect changes and skip build if no changes.
Improved the handling of variable batch file arguments.

Web Proxy:
Refactored the web proxy so that it is now implemented as an IHttpHandler rather than a web page.
The proxy is now designed to be configured into IIS to handle all VERBS for all files.